requirements.txt
alabaster==0.7.10
Babel==2.5.1
baidu-aip==2.2.0.0
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
coverage==4.4.2
crcmod==1.7
cssselect==1.0.3
Django==2.0
djangorestframework==3.7.7
docutils==0.14
et-xmlfile==1.0.1
fake-useragent==0.1.10
futures==3.1.1
gitdb==0.6.4
GitPython==0.3.6
google-api-core==0.1.4
google-auth==1.4.0
google-cloud-vision==0.30.0
googleapis-common-protos==1.5.3
grpcio==1.9.1
html2text==2018.1.9
httpie==0.9.9
idna==2.6
imagesize==0.7.1
jdcal==1.3
jieba==0.39
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
mongoengine==0.15.0
nose==1.3.7
numpy==1.14.0
openpyxl==2.5.2
oss2==2.4.0
pandas==0.22.0
parse==1.8.2
patool==1.12
Pillow==5.0.0
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyee==5.0.0
PyExecJS==1.5.1
Pygments==2.2.0
pymongo==3.6.0
pyocr==0.5
pypi-publisher==0.0.4
pyppeteer==0.0.17
pyquery==1.4.0
python-dateutil==2.6.1
python-docx==0.8.6
pytz==2017.3
requests==2.18.4
requests-html==0.9.0
rsa==3.4.2
six==1.11.0
smmap==0.9.0
snowballstemmer==1.2.1
Sphinx==1.6.6
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
tabula-py==1.0.0
urllib3==1.22
w3lib==1.19.0
Wand==0.4.4
websockets==4.0.1
xlrd==1.1.0
xlwt==1.3.0
xpdf-python==0.0.10
flask
selenium
pdfplumber
pypinyin
aip
docx
schedule
pyhanlp
elasticsearch
jsbeautifier==1.7.5
js2py